backend/requirements.txt
fastapi
uvicorn
python-dotenv
browser-use
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
requests
bcrypt==4.0.1
google-generativeai